Add authentication to get id by specific entry and delete id

+++ src/main/java/com/yash/journalApp/service/JournalEntryService.java
@@ -55,11 +55,21 @@ public class JournalEntryService {
 
         return journalEntryRepository.findById(id);
     }
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try{
+            User user = userService.findByUserName(userName);
+            removed =  user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if (removed){
+                userService.saveEntry(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch ( Exception e ){
+            System.out.println(e.getMessage());
+            throw new RuntimeException("An Error occurred while deleting the entry !" + e);
+        }
+        return removed;
     }
 
 }
